:truck: :: Move usecase
